feat: added progress effect to all loaders
