+++ src/utils/loader.ts
@@ -1,4 +1,5 @@
 import * as vscode from 'vscode'
+import { sleep } from './sleep'
 
 export async function loader(
   title: string,
@@ -18,9 +19,34 @@ export async function loader(
       title,
     },
     async (progress) => {
-      progress.report({ increment: 0 })
-      await waitFor(progress)
+      let increment = 1
+      let total = 0
+      progress.report({ increment })
+
+      const promise = waitFor(progress)
+
+      // once the process finishes, move fast towards 100%
+      let finished = false
+      let target = 0
+      const done = () => {
+        finished = true
+        target = 100
+      }
+      promise.then(done).catch(done)
+
+      // move slowly towards 100% while process not finished
+      while (!finished || total < 99) {
+        increment += (target - increment) / 50
+        total += increment
+        progress.report({ increment })
+        await sleep(50)
+      }
+
+      // finish progress
       progress.report({ increment: 100 })
+
+      // this await here is just in case the promise rejected, then this bubbles the error up, if it resolved then it just ignores it
+      await promise
     }
   )
 }
